feat: add `/revive` and `/event`

+++ src/functions/resetCheckins.ts
@@ -6,23 +6,25 @@ import log from "../utils/log";
 export default async function resetCheckins(prisma: PrismaClient, guild: Guild) {
     return new Promise<void>(async (resolve) => {
         const settings = await prisma.setting.findFirst();
-        const { checkinChannelID } = settings!;
-        // delete all messages in checkin channel
-        const channel = (await guild.channels.fetch(checkinChannelID)) as TextChannel;
+
+        if (!settings) {
+            return resolve();
+        }
+
+        const channel = (await guild.channels.fetch(settings.checkinChannelID)) as TextChannel;
         let fetched: Collection<Snowflake, Message>;
+
         do {
             const messages = await channel.messages.fetch({ limit: 100 });
             fetched = messages;
             await Promise.allSettled(messages.map((m) => m.delete()));
         } while (fetched.size >= 2);
 
-        // send check-in message
         const embed = await buildCheckinEmbed(prisma);
         await channel.send({ embeds: [embed] });
 
         const totalPlayerCount = await prisma.player.count();
 
-        // find players that are idle
         const idlePlayerCount = await prisma.player.count({
             where: {
                 isCheckedIn: false,
@@ -30,13 +32,22 @@ export default async function resetCheckins(prisma: PrismaClient, guild: Guild) 
             }
         });
 
-        // reset players that did not check in today, and reset their streak
-        const notCheckedinPlayers = await prisma.player.updateMany({
+        const playersToReset = await prisma.player.findMany({
             where: { isCheckedIn: false, checkinStreak: { gt: 0 } },
-            data: { checkinStreak: 0 }
+            select: {
+                discordID: true,
+                checkinStreak: true
+            }
         });
 
-        // reset players that have checked in today
+        const notCheckedinPlayers = await prisma.player.updateMany({
+            where: { isCheckedIn: false, checkinStreak: { gt: 0 } },
+            data: {
+                checkinStreak: 0,
+                previousStreak: playersToReset[0]?.checkinStreak ?? 0
+            }
+        });
+
         const checkedinPlayers = await prisma.player.updateMany({
             where: { isCheckedIn: true },
             data: { isCheckedIn: false }
@@ -46,6 +57,7 @@ export default async function resetCheckins(prisma: PrismaClient, guild: Guild) 
 **${checkedinPlayers.count}/${totalPlayerCount}** players checked in yesterday and have increased their progress +1
 **${notCheckedinPlayers.count}/${totalPlayerCount}** failed to check in yesterday and reset their progress back to 0
 **${idlePlayerCount}/${totalPlayerCount}** players are idle and also did not check in yesterday`;
+
         await log({ title: "Check-in Reset", content, color: "Purple" });
         return resolve();
     });
